Bumps project version to 0.4.1 to match version created by codna forge bot.
This bump should have happened before release 0.4.1. I forgot it, though,
so conda forge just came up with a version number. This commit bumps
the version so that it matches the version crated by the conda forge bot,
but the next release should use whatever version is appropriate.

The tag v0.4.1 has been applied to the commit that produced the release
0.4.1, rather than this one, so that one can track the release back to
the originating source code.